refactor: store, api 절대경로 수정
#140
