+++ client/src/App.tsx
@@ -1,9 +1,9 @@
+import { useRefreshQuery } from "@apis/auth";
+import { useErrorStore } from "@stores/useErrorStore";
+import { useUserInfo } from "@stores/useUserStore";
 import { useEffect } from "react";
+import { ErrorModal } from "@components/modal/ErrorModal";
 import { WorkSpace } from "@features/workSpace/WorkSpace";
-import { useRefreshQuery } from "./apis/auth";
-import { ErrorModal } from "./components/modal/ErrorModal";
-import { useErrorStore } from "./stores/useErrorStore";
-import { useUserInfo } from "./stores/useUserStore";
 
 const App = () => {
   const { id, name, accessToken } = useUserInfo();
