+++ client/src/features/auth/AuthButton.tsx
@@ -1,8 +1,8 @@
-import { useLogoutMutation } from "@src/apis/auth";
-import { TextButton } from "@src/components/button/textButton";
-import { Modal } from "@src/components/modal/modal";
-import { useModal } from "@src/components/modal/useModal";
-import { useCheckLogin } from "@src/stores/useUserStore";
+import { useLogoutMutation } from "@apis/auth";
+import { useCheckLogin } from "@stores/useUserStore";
+import { TextButton } from "@components/button/textButton";
+import { Modal } from "@components/modal/modal";
+import { useModal } from "@components/modal/useModal";
 import { AuthModal } from "./AuthModal";
 
 export const AuthButton = () => {
